feat: leverage new dependency features (NiceGUI 3.6, Ruff 0.14.14)
- Add app.colors global configuration for Quasar color palette
- Add ui.on_exception global handler for unhandled UI errors
- Enable RUF rules in ruff config for ruff-specific linting
- Fix all RUF violations across codebase:
  - Sort __all__ exports alphabetically
  - Use list unpacking instead of concatenation
  - Add ClassVar annotations for mutable class attributes
  - Remove unnecessary int() casts
  - Properly store asyncio.create_task references
  - Fix unused unpacked variables
  - Use raw strings for regex patterns
- Document threading model in CLAUDE.md
- Add mypy override for src.ui.app (NiceGUI 3.6 attr-defined)

+++ src/ui/graph_renderer.py
@@ -562,7 +562,7 @@ def _calculate_grid_positions(node_count: int, spacing: int = 150) -> list[tuple
     """
     import math
 
-    cols = max(1, int(math.ceil(math.sqrt(node_count))))
+    cols = max(1, math.ceil(math.sqrt(node_count)))
     positions = []
     for i in range(node_count):
         row = i // cols
